Restructure interpreters using an abstract superclass
