+++ generation/src/tests/unit/test_instantiation_optimizer.py
@@ -1,5 +1,5 @@
 from proof_generation.basic_interpreter import ExecutionPhase
-from proof_generation.instantiation_optimizer import InstantiationOptimizer
+from proof_generation.optimizing_interpreters import InstantiationOptimizer
 from proof_generation.pattern import Instantiate, MetaVar
 from proof_generation.stateful_interpreter import StatefulInterpreter
